chore(accordion-web): rename reducer file
