+++ packages/pluggableWidgets/accordion-web/src/utils/CollapsedAccordionGroupsReducer.ts
